Major stylesheet changes (read commit message)
The retina-based code that loads in the the sprite (either 'snSprite.png' or 'snSprite@2x.png') moves to its own Sass partial'_retinaMediaQuery.scss'. Read the comments at the top of this new partial to understand how things work and let me know if something's not clear...it's a tad complicated.