hide last breadcrumb elements on mobile views (see msg)
create a class called 'breadcrumb-last' and apply it to all views. it hides the page title and last &raquo; element on all single post pages when viewed on handhelds. this hides these things on the other views so set it to 'display:inline' in those spots so it's visible. apply this class to all post pages via the '_includes/post_breadcrumbs_partial.html' Jejyll include.