make sure footer headers have dotted line in all views (see msg)
place the dotted bottom border that appears on all footer headers in grunt/cssSource/_globalStyles.scss and remove it from the footer-header mixin in grunt/cssSource/_base.scss. we could put the bottom-border code on the mixin but because of how the css is built, we wouldf have to put a text shadow on the copy...we don't want to do that.