more home page construction..design & logic (see msg)
load the most recent article at the top of the home page.  more tests on
image loading by add img an image to the media query post (it's referenced
by the 'has-home-img' variable in the post template). Start using Liquid
templates to create logic that lists all the page's articles in a certain way.