add the 'jekyll:dev' task to the end of the 'push' task (see msg)
The 'push' task deploys the site to production. This task includes the 'grunt cdn' subtask, which appends the MaxCDN url to certain assets in the '_site' folder. The MaxCDN urls stay in the '_site' folder after deployment but this is the folder that the development environment runs off of. If the urls stay there, it will pull files from MaxCDN and incur costs. Add the 'jekyll:dev' task to the end of the 'push' task so jekyll rebuilds the site after everything is deployed. This will remove the CDN urls from the dev environment.