start creating separate tasks for dev site & prod site pushes (see msg)
Start creating a task that builds the site via Jekyll, creates the cache manifest, then pushes to the DEVELOPMENT site. Also start creating a task that builds the site via Jekyll with lsi stuff (custom posts), appends the MaxCDN assets to the proper elements, minifies the HTML, creates the cache manifest, then deploys it to the PRODUCTION site.